Added IsFixEditor Function to DB
for some inexplicable reason, SSMS decided to change the order of the
objects in the generated script, which makes the diff much larger than
it should be.  Honest I did nothing but add a new function.  I can find
no option to se the object order.  Lets hope this doesn't happen often.